--- src/Prism.Plugin.Logging.Common/Properties/AssemblyInfo.cs
@@ -1,3 +1,0 @@
-﻿using System.Runtime.CompilerServices;
-
-[assembly: InternalsVisibleTo("Prism.Plugin.Logging.Common.Tests, PublicKey=0024000004800000940000000602000000240000525341310004000001000100894d1945a2fa51a4225b3229c2de9da39c4e3e8a54ba70930d5b9a3fb75d379b05ebbd3ce7f930be67b906d2f7c43c8df14ea4f087252ddf184eebf0e1aedf352c0652be2b7c198ec6e6e687c73d945cccaedc70945708b729c4e91732fa52d88ff76c955ee103d32d5e99f2f70220b6d677eaed108a1e42ea1e28df546c42b3")]
