+++ tests/Prism.Plugin.Logging.AppCenter.Tests/Prism.Plugin.Logging.AppCenter.Tests.csproj
@@ -1,13 +1,15 @@
 <Project Sdk="Microsoft.NET.Sdk">
 
   <PropertyGroup>
-    <TargetFramework>netcoreapp3.1</TargetFramework>
+    <TargetFramework>net47</TargetFramework>
     <IsPackable>false</IsPackable>
-    <RootNamespace>Prism.Plugin.Logging.Common</RootNamespace>
+    <RootNamespace>Prism.Plugin.Logging</RootNamespace>
+    <SignAssembly>false</SignAssembly>
   </PropertyGroup>
 
   <ItemGroup>
     <PackageReference Include="Microsoft.NET.Test.Sdk" Version="16.7.1" />
+    <PackageReference Include="Moq" Version="4.14.5" />
     <PackageReference Include="xunit" Version="2.4.1" />
     <PackageReference Include="xunit.runner.visualstudio" Version="2.4.3">
       <PrivateAssets>all</PrivateAssets>
@@ -21,7 +23,7 @@
 
   <ItemGroup>
     <ProjectReference Include="..\..\src\Prism.Plugin.Logging.Abstractions\Prism.Plugin.Logging.Abstractions.csproj" />
-    <ProjectReference Include="..\..\src\Prism.Plugin.Logging.Common\Prism.Plugin.Logging.Common.csproj" />
+    <ProjectReference Include="..\..\src\Prism.Plugin.Logging.AppCenter\Prism.Plugin.Logging.AppCenter.csproj" />
   </ItemGroup>
 
 </Project>
