--- tests/Prism.Plugin.Logging.Common.Tests/Tests/FileSystem.cs
@@ -1,4 +1,0 @@
-﻿namespace Prism.Plugin.Logging.Common.Tests
-{
-    public class FileSystem { }
-}
