+++ tests/Prism.Plugin.Logging.Tests/Tests/FileSystem.cs
@@ -1,0 +1,4 @@
+﻿namespace Prism.Plugin.Logging.Tests
+{
+    public class FileSystem { }
+}
