+++ tests/Prism.Plugin.Logging.Tests/Tests/FileSystemTestCollection.cs
@@ -1,6 +1,6 @@
 ﻿using Xunit;
 
-namespace Prism.Plugin.Logging.Common.Tests
+namespace Prism.Plugin.Logging.Tests
 {
     [CollectionDefinition(nameof(FileSystem), DisableParallelization = true)]
     public class FileSystemTestCollection : ICollectionFixture<FileSystem>
